+++ src/components/Search/index.tsx
@@ -3,6 +3,9 @@ import MonthsAutocomplete from './MonthsAutocomplete';
 import AgencyTypeSelect from './AgencyTypeSelect';
 import AgencyAutocomplete from './AgencyAutocomplete';
 import CategorySelect from './CategorySelect';
+import ClearButton from './ClearButton';
+import SearchButton from './SearchButton';
+import Result from './Result';
 
 const Search = {
   YearsAutocomplete: YearAutocomplete,
@@ -10,6 +13,9 @@ const Search = {
   AgencyTypeSelect: AgencyTypeSelect,
   AgencyAutocomplete: AgencyAutocomplete,
   CategorySelect: CategorySelect,
+  ClearButton: ClearButton,
+  SearchButton: SearchButton,
+  Result: Result,
 };
 
 export default Search;
